Add comp flags and fix comp errors

+++ src/minishell.h
@@ -208,7 +208,7 @@ t_error		exec_one_builtin(t_cmd *cmd);
 // fds.c--------------------------------------------------------------------//
 t_error		reset_io(void);
 t_error		set_io_redirs(t_list	*redir_lst, char *hd_str);
-t_error		set_io_pipes(int child_i, size_t num_childs);
+t_error		set_io_pipes(size_t child_i, size_t num_childs);
 t_error		close_unused_fds(size_t i, size_t num_childs);
 
 // fds2.c -------------------------------------------------------------------//
